restructure setting activity
