+++ code/app/src/main/java/com/example/qrcodesfornoobs/Activity/SettingsActivity.java
@@ -1,6 +1,5 @@
-package com.example.qrcodesfornoobs;
+package com.example.qrcodesfornoobs.Activity;
 
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
@@ -8,15 +7,17 @@ import android.widget.Button;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-public class Settings extends AppCompatActivity {
+import com.example.qrcodesfornoobs.R;
+
+public class SettingsActivity extends AppCompatActivity {
     Button backButton;
-    private Intent dashboardIntent;
+    private Intent mainIntent;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.settings);
-        dashboardIntent = new Intent(this, Dashboard.class);
+        setContentView(R.layout.activity_settings);
+        mainIntent = new Intent(this, MainActivity.class);
         addListenerOnButtons();
     }
 
@@ -25,7 +26,7 @@ public class Settings extends AppCompatActivity {
         backButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                startActivity(dashboardIntent);
+                startActivity(mainIntent);
             }
         });
     }
