+++ code/app/src/main/java/com/example/qrcodesfornoobs/Dashboard.java
@@ -10,6 +10,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.qrcodesfornoobs.Activity.SettingsActivity;
 import com.example.qrcodesfornoobs.Adapter.CodeSliderAdapter;
 import com.example.qrcodesfornoobs.databinding.DashboardBinding;
 import com.google.android.material.navigation.NavigationBarView;
@@ -37,7 +38,7 @@ public class Dashboard extends AppCompatActivity {
 //        setContentView(R.layout.dashboard);
 
         profileIntent = new Intent(this, Profile.class);
-        settingsIntent = new Intent(this, Settings.class);
+        settingsIntent = new Intent(this, SettingsActivity.class);
         searchIntent = new Intent(this, Search.class);
         leaderboardIntent = new Intent(this, Leaderboard.class);
         dashboardIntent = new Intent(this, Dashboard.class);
