+++ code/app/src/main/java/com/example/qrcodesfornoobs/Fragment/DashboardFragment.java
@@ -11,7 +11,7 @@ import android.view.ViewGroup;
 
 import com.example.qrcodesfornoobs.Adapter.CodeSliderAdapter;
 import com.example.qrcodesfornoobs.Profile;
-import com.example.qrcodesfornoobs.Settings;
+import com.example.qrcodesfornoobs.Activity.SettingsActivity;
 import com.example.qrcodesfornoobs.databinding.FragmentDashboardBinding;
 
 
@@ -36,7 +36,7 @@ public class DashboardFragment extends Fragment {
         binding = FragmentDashboardBinding.inflate(getLayoutInflater());
 
         profileIntent = new Intent(getActivity(), Profile.class);
-        settingsIntent = new Intent(getActivity(), Settings.class);
+        settingsIntent = new Intent(getActivity(), SettingsActivity.class);
         setUpSliders();
         addListenerOnButtons();
